Update logic for session interruption
In testing hiding of the info view when
going between background and foreground
it became apparent that the logic for
pausing the ARSession was not working
as desired. When coming to the foreground
it would seem that the session was paused
and could not be resumed. This addresses
the issue by not pausing